Support wildcards in newsgroup names
I was going to get a list of all newsgroups using LIST NEWSGROUPS,
then get the program to locally pick out only ones which match the
non-wildcard beginning of the string (a wildcard is only allowed at
the end of the string), but it turns out that RFC3977 section 4
already has this functionality built in.  Fantastic!