Revert "Make quick script to fix broken hashes"
This reverts commit f1743728970d626ce77522ef414b035f94761620.

This needs a bit more tweaking; as it currently stands, import.py
deletes duplicates, meaning in this instance that it deletes messages
that have the correct ID.  I'll add a "don't delete duplicates"
option first, or better yet, a "do delete duplicates" option, with
not deleting them being the default behaviour.  In the meantime,
*don't* use import.py or fix-hashes.sh as recommended in the last
commit.